Moved cmd/basecli/commands into client/commands
