+++ client/commands/auto/cmd.go
@@ -1,4 +1,4 @@
-package commands
+package auto
 
 import (
 	"os"
